+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New #92: Add `canSendPaidMedia` parameter to `ChatFullInfo` type.
 - New #93: Add `paidMedia` parameter to `Message` and `ExternalReplyInfo` types, add `PaidMediaInfo`,
   `PaidMediaPreview`, `PaidMediaPhoto`, `PaidMediaVideo` types.
+- New #94: Add `invoicePayload` parameter to `TransactionPartnerUser` type.
 
 ## 0.2.0 June 29, 2024